create invite to team helper
